+++ Frends.Pdf.Create/Frends.Pdf.Create/Definitions/Result.cs
@@ -9,17 +9,17 @@ public class Result
     /// Indicates if the operation was successful.
     /// </summary>
     /// <example>true</example>
-    public bool Success { get; private set; }
+    public bool Success { get; init; }
 
     /// <summary>
     /// Name of the file which was created.
     /// </summary>
     /// <example>C:\tmp\example_file.pdf</example>
-    public string FileName { get; private set; }
+    public string FileName { get; init; }
 
-    internal Result(bool success, string fileName)
-    {
-        Success = success;
-        FileName = fileName;
-    }
+    /// <summary>
+    /// Error details. Null when Success is true.
+    /// </summary>
+    /// <example>null</example>
+    public Error Error { get; init; }
 }